+++ test/com/lynxanalytics/biggraph/spark_util/HybridRDDTest.scala
@@ -3,6 +3,7 @@ package com.lynxanalytics.biggraph.spark_util
 import org.scalatest.FunSuite
 import org.apache.spark.HashPartitioner
 import org.apache.spark.rdd.RDD
+import com.lynxanalytics.biggraph.graph_api.RuntimeContext
 import com.lynxanalytics.biggraph.TestSparkContext
 
 class HybridRDDTest extends FunSuite with TestSparkContext {
@@ -22,24 +23,41 @@ class HybridRDDTest extends FunSuite with TestSparkContext {
     }
 
     import Implicits._
+    val partitioner = new HashPartitioner(10)
     val sourceRDD = sparkContext.parallelize(localSource, 10)
-    val lookupRDD = sparkContext.parallelize(localLookup).sortUnique(new HashPartitioner(10))
+    val lookupRDD = sparkContext.parallelize(localLookup).sortUnique(partitioner)
 
-    checkGood(HybridRDD(sourceRDD, 2000).lookup(lookupRDD))
-    checkGood(HybridRDD(sourceRDD, 200).lookup(lookupRDD))
-    checkGood(HybridRDD(sourceRDD, 20).lookup(lookupRDD))
-    checkGood(HybridRDD(sourceRDD, 2).lookup(lookupRDD))
-    checkGood(HybridRDD(sourceRDD, 0).lookup(lookupRDD))
+    implicit val rc = RuntimeContext(sparkContext, null, null, null, null)
+    checkGood(HybridRDD(sourceRDD, partitioner, even = true, 2000).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = true, 200).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = true, 20).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = true, 2).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = true, 0).lookup(lookupRDD))
+
+    checkGood(HybridRDD(sourceRDD, partitioner, even = false, 2000).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = false, 200).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = false, 20).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = false, 2).lookup(lookupRDD))
+    checkGood(HybridRDD(sourceRDD, partitioner, even = false, 0).lookup(lookupRDD))
   }
 
   test("lookup on empty RDD") {
     import Implicits._
+    val partitioner = new HashPartitioner(1)
     val sourceRDD = sparkContext.emptyRDD[(Int, Long)]
-    val lookupRDD = sparkContext.emptyRDD[(Int, Double)].sortUnique(new HashPartitioner(1))
-    assert(HybridRDD(sourceRDD, 2000).lookup(lookupRDD).collect.isEmpty)
-    assert(HybridRDD(sourceRDD, 200).lookup(lookupRDD).collect.isEmpty)
-    assert(HybridRDD(sourceRDD, 20).lookup(lookupRDD).collect.isEmpty)
-    assert(HybridRDD(sourceRDD, 2).lookup(lookupRDD).collect.isEmpty)
-    assert(HybridRDD(sourceRDD, 0).lookup(lookupRDD).collect.isEmpty)
+    val lookupRDD = sparkContext.emptyRDD[(Int, Double)].sortUnique(partitioner)
+    implicit val rc = RuntimeContext(sparkContext, null, null, null, null)
+
+    assert(HybridRDD(sourceRDD, partitioner, even = true, 2000).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = true, 200).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = true, 20).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = true, 2).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = true, 0).lookup(lookupRDD).collect.isEmpty)
+
+    assert(HybridRDD(sourceRDD, partitioner, even = false, 2000).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = false, 200).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = false, 20).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = false, 2).lookup(lookupRDD).collect.isEmpty)
+    assert(HybridRDD(sourceRDD, partitioner, even = false, 0).lookup(lookupRDD).collect.isEmpty)
   }
 }